new dynamic size bitmap
